Removes all HTTPS stuff for dev server
Too complicated, probably not needed. Problem is that Shadow CLJS uses
websockets for REPL and hot reloading. The TLS on websockets doesn't
work through the same mechanism in iOS as HTTPS does, meaning that even
when HTTPS works, websockets don't. If really needed, clues for a
solution may be found here:

https://github.com/eamonwhiter73/IOSObjCWebSockets

And:

https://developer.apple.com/documentation/network/implementing_netcat_with_network_framework